pyproject.toml
[tool.poetry]
name = "fakerz"
version = "0.1.0"
description = "Fake API server"
authors = ["Mohamed Nesredin"]
license = "MIT"

[tool.poetry.dependencies]
python = "3.8.2"
fastapi = {extras = ["all"], version = "^0.54.1"}
loguru = "^0.4.1"
pendulum = "^2.1.0"
Faker = "^4.0.1"
httpx = "^0.12.1"
gunicorn = "^20.0.4"
pyjwt = "^1.7.1"
tortoise-orm = "^0.16.7"
python-dotenv = "^0.13.0"
asyncpg = "^0.20.1"
confusable_homoglyphs = "^3.2.0"

[tool.poetry.dev-dependencies]
bandit = "^1.6.2"
black = "^19.10b0"
pytest = "^5.4.1"
coverage = "^5.1"
isort = "^4.3.21"
mypy = "^0.770"
radon = "^4.1.0"
pytest-sugar = "^0.9.2"
pytest-cov = "^2.8.1"
safety = "^1.8.7"
mkdocs = "^1.1"
mkdocs-material = "^5.1.1"

[build-system]
requires = ["poetry>=0.12"]
build-backend = "poetry.masonry.api"
